+++ includes/inspector/index.php
@@ -75,10 +75,95 @@ add_action('wp_footer', function() {
             require_once 'panels/main.php';
             require_once 'panels/errors.php';
             require_once 'panels/nav.php';
+            require_once 'panels/basset-object.php';
             do_action('basset/inspector/panels');
             ?>
         </div><!-- / #basset_theme_inspector -->
         <?php
     }
 });
+
+function basset_print_property_list($id, $object) {
+    if (isset($object)) {
+        $level = 0;
+        ?>
+        <table id="<?=$id?>" class="property_list">
+            <thead>
+                <tr>
+                    <th>Key</th>
+                    <th>Type</th>
+                    <th>Value</th>
+                </tr>
+            </thead>
+            <tbody>
+                <?php
+                foreach($object as $property => $value) {
+                    basset_print_property_list_row($property, $value, $id, $level, $id);
+                }
+                ?>
+            </tbody>
+        </table><!-- /.property_list -->
+        <?php
+    }
+}
+
+function basset_print_property_list_row($property, $value, $parent = '', $level = 0, $table_id = '') {
+    if (is_array($value) || is_object($value)) {
+        $is_collection = true;
+        $has_children = "true"; // for data- attr.
+        $icon = "dashicons-arrow-right";
+    } else {
+        $is_collection = false;
+        $has_children = "false"; // for data- attr.
+        $icon = "";
+    }
+    if (isset($parent)) {
+        $parent = "data-parent='$table_id-$parent'";
+    }
+    ?>
+    <tr id="<?= $table_id . '-' . $property ?>" class="property-summary-row" <?=$parent?> data-level="<?=$level?>" data-is-collection="<?=$has_children?>">
+        <? if (is_int($property)) { $property = "Item: $property"; } ?>
+        <td class="basset-property-list-key"><span class="dashicons <?=$icon?>"></span> <?php print $property ?></td>
+        <td class="basset-property-list-type"><?php print gettype($value) ?></td>
+        <td class="basset-property-list-value"><?php print basset_property_summary($value) ?></td>
+    </tr>
+    <?
+    if ($is_collection && count($value)) {
+        $parent = $property;
+        $level++;
+        foreach($value as $property => $value) {
+            basset_print_property_list_row($property, $value, $parent, $level, $table_id);
+        }
+    }
+}
+
+function basset_property_summary($value) {
+    $type = gettype($value);
+    switch ($type) {
+        case "boolean":
+            if ($value) {
+                $checked = 'checked="checked"';
+            } else {
+                $checked = false;
+            }
+            return "<input type='checkbox' $checked>";
+        case "integer":
+            return "Number (int)";
+        case "double":
+            return "Number (double)";
+        case "string":
+            // @TODO: Check value length - send back short strings or truncate.
+            return $value;
+        case "array":
+            return count($value) . " Items";
+        case "object":
+            return "{...}";
+        case "resource":
+            return "Resource";
+        case "NULL":
+            return "null";
+        default:
+            return "Unknown";
+    }
+}
 ?>